INC-119 fix: getting notifications logic refactored to custom hook useNotifications
