+++ src/components/ui/header/Header.tsx
@@ -1,15 +1,10 @@
-import { useCallback, useEffect, useState } from 'react'
-
 import { useTranslation } from '@/common/hooks/useTranslation'
 import { LangSelect } from '@/components/langSelect/LangSelect'
 import { NotificationsDropDown } from '@/components/notificationDropdown/notificationsDropDown'
+import { useNotifications } from '@/components/notificationDropdown/useNotifications'
 import { Button } from '@/components/ui/button'
 import { Typography } from '@/components/ui/typography'
 import { useMeQuery } from '@/services/auth'
-import { useGetNotificationsQuery } from '@/services/notifications/notificationsService'
-import { NotificationType } from '@/services/notifications/notificationsService.types'
-import { useSocket } from '@/services/websockets/socket'
-import { getCookie } from 'cookies-next/client'
 import Link from 'next/link'
 
 import s from './header.module.scss'
@@ -24,29 +19,7 @@ export const Header = () => {
 
   const { data } = useMeQuery()
 
-  const { data: serverNotifications } = useGetNotificationsQuery({
-    pageSize: 100,
-    sortDirection: 'desc',
-  })
-
-  const [notifications, setNotifications] = useState<NotificationType[]>(
-    serverNotifications?.items ?? []
-  )
-
-  useEffect(() => {
-    if (serverNotifications?.items) {
-      setNotifications(serverNotifications?.items)
-    }
-  }, [serverNotifications?.items])
-  const handleNewNotification = useCallback(
-    (notification: NotificationType) => {
-      setNotifications(prev => [notification, ...prev])
-    },
-    [setNotifications]
-  )
-  const accessToken = getCookie('accessToken') as string
-
-  useSocket(accessToken, handleNewNotification)
+  const { notifications } = useNotifications()
 
   const isMyProfile = !!data
